+++ plugins/azure-wakamiti-plugin/src/main/java/es/iti/wakamiti/azure/AzureConfigContributor.java
@@ -3,8 +3,8 @@ package es.iti.wakamiti.azure;
 import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.WakamitiException;
 import es.iti.wakamiti.api.extensions.ConfigContributor;
-import imconfig.Configuration;
-import imconfig.Configurer;
+import es.iti.wakamiti.api.imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configurer;
 
 import java.util.function.BiConsumer;