+++ plugins/email-wakamiti-plugin/src/main/java/es/iti/wakamiti/email/EmailConfigContributor.java
@@ -3,8 +3,8 @@ package es.iti.wakamiti.email;
 
 import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.extensions.ConfigContributor;
-import imconfig.Configuration;
-import imconfig.Configurer;
+import es.iti.wakamiti.api.imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configurer;
 
 
 @Extension(provider =  "es.iti.wakamiti", name = "email-step-config", version = "2.5",