+++ plugins/file-uploader-wakamiti-plugin/src/test/java/es/iti/wakamiti/fileuploader/it/StandardOutputFilesUploaderTest.java
@@ -8,8 +8,8 @@ package es.iti.wakamiti.fileuploader.it;
 
 import es.iti.wakamiti.fileuploader.MockFtpServer;
 import es.iti.wakamiti.junit.WakamitiJUnitRunner;
-import imconfig.AnnotatedConfiguration;
-import imconfig.Property;
+import es.iti.wakamiti.api.imconfig.AnnotatedConfiguration;
+import es.iti.wakamiti.api.imconfig.Property;
 import org.apache.ftpserver.ftplet.FtpException;
 import org.junit.AfterClass;
 import org.junit.BeforeClass;