+++ wakamiti-plugin-maven-archetype/src/main/resources/archetype-resources/src/main/java/es/iti/wakamiti/__pluginId__/__PluginId__ConfigContributor.java
@@ -8,8 +8,8 @@ package ${package};
 
 
 
-import imconfig.Configuration;
-import imconfig.Configurer;
+import es.iti.wakamiti.api.imconfig.Configuration;
+import es.iti.wakamiti.api.imconfig.Configurer;
 import es.iti.commons.jext.Extension;
 import es.iti.wakamiti.api.extensions.ConfigContributor;