+++ wakamiti-engine/CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+
 ## [unreleased]
 
 ### Added
@@ -15,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 - The junit wakamiti configuration is modified to set the default output path of the current project.
+- Reset classpath on every run of `wakamiti-maven-plugin` verify mojo.
 
 
 ## [2.5.0] - 2024-04-16